Fixed 6-size.c

+++ 0x00-hello_world/6-size.c
@@ -18,10 +18,10 @@ int main(void)
 	ll = 5;
 	f = 5;
 
-	printf("Size of a char: %lu byte(s) \n", (unsigned long)sizeof(c));
-	printf("Size of a int: %lu byte(s) \n", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s) \n", (unsigned long)sizeof(l));
-	printf("Size of a long long: %lu byte(s) \n", (unsigned long)sizeof(ll));
-	printf("Size of a float: %lu byte(s) \n", (unsigned long)sizeof(f));
+	printf("Size of a char: %d byte(s) \n", sizeof(c));
+	printf("Size of a int: %d byte(s) \n", sizeof(i));
+	printf("Size of a long int: %d byte(s) \n", sizeof(l));
+	printf("Size of a long long: %d byte(s) \n", sizeof(ll));
+	printf("Size of a float: %d byte(s) \n", sizeof(f));
 	return (0);
 }
